ClassifierApplication/requirements.txt
imutils==0.5.3
matplotlib==3.3.0
mysql_connector_repackaged==0.3.1
opencv_python>4.0.0
pandas>1.0.0
readchar==2.0.1
tqdm==4.50.2
